改正了顽固的bug
修复了战斗结束后，军营士兵显示并没有因为战斗消耗士兵而发生变化的问题

+++ Classes/Managers/BuildingManager.cpp
@@ -1042,13 +1042,13 @@ void BuildingManager::restoreArmyCampTroopDisplays()
         
         for (int i = 0; i < count; ++i)
         {
-            if (armyCampIndex >= armyCamps.size())
+            if (armyCampIndex >= static_cast<int>(armyCamps.size()))
                 armyCampIndex = 0;
             
             armyCamps[armyCampIndex]->addTroopDisplay(unitType);
             
             if ((i + 1) % 5 == 0)
-                armyCampIndex = (armyCampIndex + 1) % armyCamps.size();
+                armyCampIndex = (armyCampIndex + 1) % static_cast<int>(armyCamps.size());
         }
     }
 }
